% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_max_sum_ratio_plot.R
\name{stat_max_sum_ratio_plot}
\alias{stat_max_sum_ratio_plot}
\title{Title}
\usage{
stat_max_sum_ratio_plot(
  mapping = NULL,
  data = NULL,
  geom = "step",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  p = 4,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
