% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_zipf.R
\name{stat_zipf}
\alias{stat_zipf}
\title{Title}
\usage{
stat_zipf(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
